+++ crates/query-engine/translation/tests/tests.rs
@@ -190,6 +190,14 @@ async fn it_aggregate_function_albums() {
 }
 
 #[tokio::test]
+async fn aggregate_limit_offset_order_by() {
+    let result = common::test_translation("aggregate_limit_offset_order_by")
+        .await
+        .unwrap();
+    insta::assert_snapshot!(result);
+}
+
+#[tokio::test]
 async fn it_simple_array_relationship() {
     let result = common::test_translation("simple_array_relationship")
         .await